requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration

openai

# Anthropic Integration


# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

pydantic
streamlit


grandalf

langchain-community

pypdf

bs4